Improve package command slightly. Still needs testing on windows.

+++ conanfile.py
@@ -38,22 +38,18 @@ class GeographicLibConan(ConanFile):
         self.run('cmake --build build %s' % cmake.build_config)
 
     def package(self):
-        self.copy(pattern="*.hpp", dst="include", src="distrib/include")
-        self.copy(pattern="*/Config.h", dst="include", src="build/include")
-        self.copy(pattern="*.so", dst="lib", src="build/src", keep_path=False)
-        self.copy(pattern="*.a", dst="lib", src="build/src", keep_path=False)
-
-        build_dir = "build/src"
+        if self.options.shared:
+            build_dir = "build"
+        else:
+            build_dir = "build/lib"
 
         if self.settings.os == "Windows":
             if self.options.shared:
                 self.copy(pattern="*.dll", dst="bin", src=build_dir, keep_path=False)
-                # build_dir = os.path.join(self.ZIP_FOLDER_NAME, "_build/lib")
                 self.copy(pattern="*GeographicLibd.lib", dst="lib", src=build_dir, keep_path=False)
                 self.copy(pattern="*GeographicLib.lib", dst="lib", src=build_dir, keep_path=False)
                 self.copy(pattern="*GeographicLib.dll.a", dst="lib", src=build_dir, keep_path=False)
             else:
-                # build_dir = os.path.join(self.ZIP_FOLDER_NAME, "_build/lib")
                 if self.settings.os == "Windows":
                     # MinGW
                     self.copy(pattern="libGeographicLibstaticd.a", dst="lib", src=build_dir, keep_path=False)
@@ -66,19 +62,21 @@ class GeographicLibConan(ConanFile):
                 suffix = "d" if self.settings.build_type == "Debug" else ""
                 if self.settings.compiler == "Visual Studio":
                     current_lib = os.path.join(lib_path, "GeographicLibstatic%s.lib" % suffix)
-                    os.rename(current_lib, os.path.join(lib_path, "GeographicLib%s.lib" % suffix))
+                    if os.path.isfile(current_lib):
+                        os.rename(current_lib, os.path.join(lib_path, "GeographicLib%s.lib" % suffix))
                 elif self.settings.compiler == "gcc":
                     current_lib = os.path.join(lib_path, "libGeographicLibstatic.a")
-                    os.rename(current_lib, os.path.join(lib_path, "libGeographicLib.a"))
-        else:
-            if self.options.shared:
-                if self.settings.os == "Macos":
-                    self.copy(pattern="*.dylib", dst="lib", src=build_dir, keep_path=False)
-                else:
-                    self.copy(pattern="*.so*", dst="lib", src=build_dir, keep_path=False)
-            else:
-                self.copy(pattern="*.a", dst="lib", src=build_dir, keep_path=False)
+                    if os.path.isfile(current_lib):
+                        os.rename(current_lib, os.path.join(lib_path, "libGeographicLib.a"))
 
+        self.copy(pattern="*.hpp", dst="include", src="distrib/include")
+        self.copy(pattern="*/Config.h", dst="include", src="build/include")
+        self.copy(pattern="*.so", dst="lib", src=build_dir, keep_path=False)
+        self.copy(pattern="*.so.*", dst="lib", src=build_dir, keep_path=False, symlinks=True)
+        self.copy(pattern="*.a", dst="lib", src=build_dir, keep_path=False)
+        self.copy(pattern="*.dylib*", dst="lib", src=build_dir, keep_path=False)
+        self.copy(pattern="*.lib", dst="lib", src=build_dir, keep_path=False)
+        self.copy(pattern="*.dll", dst="bin", src=build_dir, keep_path=False)
 
 
     def package_info(self):
